+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 * Changelog created.
 * Fixed some broken ExDoc links in `OSC.Types.Blob`.
+* FIxed nulls in doctests in `OSC.Types.String`.